# ENH: linalg: speed up _compute_lwork by avoiding numpy constructs