--- a/scipy/linalg/lapack.py
+++ b/scipy/linalg/lapack.py
@@ -750,6 +750,9 @@ def get_lapack_funcs(names, arrays=(), dtype=None):
                       "flapack", "clapack", _lapack_alias)
 
 
+_int32_max = _np.iinfo(_np.int32).max
+
+
 def _compute_lwork(routine, *args, **kwargs):
     """
     Round floating-point lwork returned by lapack to integer.
@@ -771,27 +774,31 @@ def _compute_lwork(routine, *args, **kwargs):
     32000
 
     """
-    wi = routine(*args, **kwargs)
-    if len(wi) < 2:
-        raise ValueError('')
-    info = wi[-1]
-    if info != 0:
-        raise ValueError("Internal work array size computation failed: "
-                         "%d" % (info,))
-
-    lwork = [w.real for w in wi[:-1]]
-
     dtype = getattr(routine, 'dtype', None)
+    ret = routine(*args, **kwargs)
+    if ret[-1] != 0:
+        raise ValueError("Internal work array size computation failed: "
+                         "%d" % (ret[-1],))
+
+    if len(ret) == 2:
+        return _check_work_float(ret[0].real, dtype)
+    else:
+        return tuple(_check_work_float(x.real, dtype) for x in ret[:-1])
+
+
+def _check_work_float(value, dtype):
+    """
+    Convert LAPACK-returned work array size float to integer,
+    carefully for single-precision types.
+    """
+
     if dtype == _np.float32 or dtype == _np.complex64:
         # Single-precision routine -- take next fp value to work
         # around possible truncation in LAPACK code
-        lwork = _np.nextafter(lwork, _np.inf, dtype=_np.float32)
+        value = _np.nextafter(value, _np.inf, dtype=_np.float32)
 
-    lwork = _np.array(lwork, _np.int64)
-    if _np.any(_np.logical_or(lwork < 0, lwork > _np.iinfo(_np.int32).max)):
+    value = int(value)
+    if value < 0 or value > _int32_max:
         raise ValueError("Too large work array required -- computation cannot "
                          "be performed with standard 32-bit LAPACK.")
-    lwork = lwork.astype(_np.int32)
-    if lwork.size == 1:
-        return lwork[0]
-    return lwork
+    return value