The _compute_lwork routine uses numpy constructs to handle cases where
also rwork/iwork is returned. Use Python ones instead to speed it up.

The micro-optimization can matter for small arrays:
```
# before
In [2]: %timeit scipy.linalg.hessenberg(np.eye(3))                                                                                     
75.5 µs ± 303 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
# after
In [2]: %timeit scipy.linalg.hessenberg(np.eye(3))                                                                                     
49.8 µs ± 1.09 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```
Line profile afterward:
```
Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  1206                                           def hessenberg(a, calc_q=False, overwrite_a=False, check_finite=True):
...
  1252         1         27.0     27.0     16.7      a1 = _asarray_validated(a, check_finite=check_finite)
  1253         1          2.0      2.0      1.2      if len(a1.shape) != 2 or (a1.shape[0] != a1.shape[1]):
  1254                                                   raise ValueError('expected square matrix')
  1255         1          3.0      3.0      1.9      overwrite_a = overwrite_a or (_datacopied(a1, a))
  1256                                           
  1257                                               # if 2x2 or smaller: already in Hessenberg
  1258         1          1.0      1.0      0.6      if a1.shape[0] <= 2:
  1259                                                   if calc_q:
  1260                                                       return a1, numpy.eye(a1.shape[0])
  1261                                                   return a1
  1262                                           
  1263         1          1.0      1.0      0.6      gehrd, gebal, gehrd_lwork = get_lapack_funcs(('gehrd', 'gebal',
  1264         1          8.0      8.0      4.9                                                    'gehrd_lwork'), (a1,))
  1265         1         19.0     19.0     11.7      ba, lo, hi, pivscale, info = gebal(a1, permute=0, overwrite_a=overwrite_a)
  1266         1          3.0      3.0      1.9      _check_info(info, 'gebal (hessenberg)', positive=False)
  1267         1          2.0      2.0      1.2      n = len(a1)
  1268                                           
  1269         1         19.0     19.0     11.7      lwork = _compute_lwork(gehrd_lwork, ba.shape[0], lo=lo, hi=hi)
  1270                                           
  1271         1          9.0      9.0      5.6      hq, tau, info = gehrd(ba, lo=lo, hi=hi, lwork=lwork, overwrite_a=1)
  1272         1          2.0      2.0      1.2      _check_info(info, 'gehrd (hessenberg)', positive=False)
  1273         1         64.0     64.0     39.5      h = numpy.triu(hq, -1)
  1274         1          1.0      1.0      0.6      if not calc_q:
  1275         1          1.0      1.0      0.6          return h
```